`refactor`(models): models modularizado em pacote
Código de `models.py` refatorado e transformado em um pacote dentro de
`app` para melhor manutenabilidade e legibilidade.

BREAKING CHANGE: Versões de migration resetadas.
